refactor: update fileUploadType filename
