+++ .github/whitelist-capital-letters.txt
@@ -39,7 +39,6 @@
 ./libs/common/src/misc/serviceUtils.ts
 ./libs/common/src/misc/serviceUtils.spec.ts
 ./libs/common/src/enums/kdfType.ts
-./libs/common/src/enums/fileUploadType.ts
 ./libs/common/src/enums/linkedIdType.ts
 ./libs/common/src/enums/htmlStorageLocation.ts
 ./libs/common/src/enums/notificationType.ts
