+++ libs/common/src/abstractions/file-upload/file-upload.service.ts
@@ -1,4 +1,4 @@
-import { FileUploadType } from "../../enums/fileUploadType";
+import { FileUploadType } from "../../enums/file-upload-type";
 import { EncArrayBuffer } from "../../models/domain/enc-array-buffer";
 import { EncString } from "../../models/domain/enc-string";
