+++ libs/common/src/services/file-upload/file-upload.service.ts
@@ -3,7 +3,7 @@ import {
   FileUploadService as FileUploadServiceAbstraction,
 } from "../../abstractions/file-upload/file-upload.service";
 import { LogService } from "../../abstractions/log.service";
-import { FileUploadType } from "../../enums/fileUploadType";
+import { FileUploadType } from "../../enums/file-upload-type";
 import { EncArrayBuffer } from "../../models/domain/enc-array-buffer";
 import { EncString } from "../../models/domain/enc-string";
 import { AzureFileUploadService } from "../azureFileUpload.service";
